\name{rndLambdaF}
\alias{rndLambdaF}
\title{
  Principal Component Statistics Based on Randomization
}
\description{
  TODO: Stuff goes here...
}
\usage{
rndLambdaF(data, B = 1000, alpha = 0.05, ...)
}
\arguments{
  \item{data}{TODO}
  \item{B}{TODO}
  \item{alpha}{TODO}
  \item{\dots}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
\references{
  TODO
}
\author{
  Kevin R. Coombes <krc@silicovore.com>, Min Wang <wang.1807@osu.edu>.
}
\seealso{
}
\examples{
}
\keyword{models}
