\name{spca-data}
\alias{spca-data}
\alias{spca}
\docType{data}
\title{Sample PCA Dataset}
\description{
  This data set contains an object of the class \code{link{SamplePCA}}.
  This object results from performing a principal components analysis on
  a simulated data set where the true 
}
\usage{
data(spca)
}
\format{A \code{SamplePCA} object based on a simulated data matrix with
  204 rows and 14 columns, with true "principal component dimension"
  equal to one. That is, there should be one significant principal
  component.} 
\source{
  Simulations are described in detail in the \code{Thresher} package,
  which depends on the \code{PCDimension} package.
}
\seealso{
  The ClassDiscovery package contains the \code{\link{SamplePCA}}
  class and functions.
}
\keyword{datasets}

